+++ x/auth/ante/fee_test.go
@@ -66,7 +66,7 @@ func (s *AnteTestSuite) TestEnsureMempoolFees() {
 	// msg and signatures
 	msg := testdata.NewTestMsg(accs[0].acc.GetAddress())
 	feeAmount := testdata.NewTestFeeAmount()
-	gasLimit := testdata.NewTestGasLimit()
+	gasLimit := uint64(15)
 	s.Require().NoError(s.txBuilder.SetMsgs(msg))
 	s.txBuilder.SetFeeAmount(feeAmount)
 	s.txBuilder.SetGasLimit(gasLimit)
@@ -76,7 +76,7 @@ func (s *AnteTestSuite) TestEnsureMempoolFees() {
 	s.Require().NoError(err)
 
 	// Set high gas price so standard test fee fails
-	atomPrice := sdk.NewDecCoinFromDec("atom", math.LegacyNewDec(20))
+	atomPrice := sdk.NewDecCoinFromDec("atom", sdk.NewDec(20))
 	highGasPrice := []sdk.DecCoin{atomPrice}
 	s.ctx = s.ctx.WithMinGasPrices(highGasPrice)
 
@@ -108,9 +108,9 @@ func (s *AnteTestSuite) TestEnsureMempoolFees() {
 
 	newCtx, err := antehandler(s.ctx, tx, false)
 	s.Require().Nil(err, "Decorator should not have errored on fee higher than local gasPrice")
-	// Priority is the smallest amount in any denom. Since we have only 1 fee
-	// of 150atom, the priority here is 150.
-	s.Require().Equal(feeAmount.AmountOf("atom").Int64(), newCtx.Priority())
+	// Priority is the smallest gas price amount in any denom. Since we have only 1 gas price
+	// of 10atom, the priority here is 10.
+	s.Require().Equal(int64(10), newCtx.Priority())
 }
 
 func (s *AnteTestSuite) TestDeductFees() {